Move libqtile.window into x11 backend, creating dedicated window.py
This is an ugly but necessary step to move x11-specific code into the
x11 backend. I played around with merging xcbq.Window and window.Window
but it's incredibly fiddly and does not provide much benefit (beside
neater code). As long as code elsewhere outside of the backend accesses
exclusively the top-level window, abstracting the interface and
creating the equivalent class for Wayland is doable.

As a temporary change, the docs reference to window.Window points to the
x11 backend's implementation. Once the wayland Window class has matched
all of the required cmd_ methods and the base Window has been updated,
the docs can point to the base class.
