+++ test/test_manager.py
@@ -38,8 +38,7 @@ import libqtile.confreader
 import libqtile.hook
 import libqtile.layout
 import libqtile.widget
-import libqtile.window
-from libqtile.backend.x11 import xcbq
+from libqtile.backend.x11 import window, xcbq
 from libqtile.command.client import SelectError
 from libqtile.command.interface import CommandError, CommandException
 from libqtile.config import Match
@@ -468,7 +467,7 @@ def test_change_state_via_message(manager):
     window_info = manager.c.window.info()
     conn = xcbq.Connection(manager.display)
 
-    data = xcffib.xproto.ClientMessageData.synthetic([libqtile.window.IconicState, 0, 0, 0, 0], "IIIII")
+    data = xcffib.xproto.ClientMessageData.synthetic([window.IconicState, 0, 0, 0, 0], "IIIII")
     ev = xcffib.xproto.ClientMessageEvent.synthetic(
         32, window_info["id"], conn.atoms['WM_CHANGE_STATE'], data
     )
@@ -476,7 +475,7 @@ def test_change_state_via_message(manager):
     conn.xsync()
     assert manager.c.window.info()["minimized"]
 
-    data = xcffib.xproto.ClientMessageData.synthetic([libqtile.window.NormalState, 0, 0, 0, 0], "IIIII")
+    data = xcffib.xproto.ClientMessageData.synthetic([window.NormalState, 0, 0, 0, 0], "IIIII")
     ev = xcffib.xproto.ClientMessageEvent.synthetic(
         32, window_info["id"], conn.atoms['WM_CHANGE_STATE'], data
     )
@@ -1540,8 +1539,8 @@ def test_strut_handling(manager):
         assert manager.c.window.info()['x'] == 820
         assert manager.c.window.info()['y'] == 0
     finally:
-        for window in w:
-            window.kill_client()
+        for win in w:
+            win.kill_client()
         conn.finalize()
 
     test_initial_state()
